refactor : update Promotion Ui to depend on the interface insted of the implementation .
